Disable length-zero-no-unit stylelint option
Stylelint also fix-breaks calc() functions as it removes units of values that need it (see https://github.com/stylelint/stylelint/issues/3037). It also auto-fixes it (and hence breaks it then): https://github.com/stylelint/stylelint/issues/2643, so we have to disable the whole rule for now. In order to be able to add comments to the stylelint config I moved it to a .js file.