Remove ENSResolver from context, register Resolver in saga
Note that the UserResolver requires an initialized network client in its
constructor. Therefore we can't register it without access to the network client
from the context.

If we want to avoid this pattern, then we can pass in the instance when using
the methods. Maybe there is a third option.