Refactor colony roles fetching
Changed the getColonyRoles query to first reduce role set events to a
mapping of address to mapping of role to boolean. This is then filtered
to find the founder - the most recent address with all roles that the
OldRoles setFounderRole function sets. It is then filtered to find all
addresses with the admin role.