Fix colony create wizard transform memoization
When setting the token icon, an object was passed to the next stage of
the form. This was later recreated as another object, which caused
memoized values to be recomputed. Since it was used in the asyncFunction
transform memoization, this could cause the asyncFunction to be
recomputed between submit and success actions, which would prevent the
success action from correctly triggering the onSuccess function of the
ActionForm.