Replace the use of blocksBack with fromBlock=1
We were previously only fetching event logs up to a certain number of
blocks ago. This was due to performance reasons and nodes taking a long
time to respond if we include all blocks.

The way in which nodes index events has apparently been improved now (at
least for Infura), and so we can query for all blocks! This fixed weird
things happening for colonies older than the blocksBack value.