Gracefully handle tx wait errors
When in a saga and you need to wait for it to be finished to do UI responses, instead of doing:
```
yield takeFrom(txChannel, ActionTypes.TRANSACTION_SUCCEEDED);
```

you do now
```
yield waitForTxResult(txChannel);;
```

and it will handle all errors gracefully for you. When there are more than one transaction channels you'll have to wait for all of their results, otherwise some errors might not get handled.